+++ src/main/java/com/axperty/stackedblocksfarmersdelight/StackedBlocksFarmersDelight.java
@@ -1,0 +1,56 @@
+package com.axperty.stackedblocksfarmersdelight;
+
+import com.axperty.stackedblocksfarmersdelight.registry.CreativeTabRegistry;
+import com.axperty.stackedblocksfarmersdelight.registry.BlockRegistry;
+import com.axperty.stackedblocksfarmersdelight.registry.ItemRegistry;
+import org.slf4j.Logger;
+
+import com.mojang.logging.LogUtils;
+
+import net.neoforged.api.distmarker.Dist;
+import net.neoforged.bus.api.IEventBus;
+import net.neoforged.bus.api.SubscribeEvent;
+import net.neoforged.fml.ModContainer;
+import net.neoforged.fml.common.EventBusSubscriber;
+import net.neoforged.fml.common.Mod;
+import net.neoforged.fml.event.lifecycle.FMLClientSetupEvent;
+import net.neoforged.fml.event.lifecycle.FMLCommonSetupEvent;
+import net.neoforged.neoforge.common.NeoForge;
+import net.neoforged.neoforge.event.server.ServerStartingEvent;
+
+@Mod(StackedBlocksFarmersDelight.MOD_ID)
+public class StackedBlocksFarmersDelight
+{
+    public static final String MOD_ID = "stackedblocksfarmersdelight";
+    private static final Logger LOGGER = LogUtils.getLogger();
+
+    public StackedBlocksFarmersDelight(IEventBus modEventBus, ModContainer modContainer)
+    {
+        modEventBus.addListener(this::commonSetup);
+        BlockRegistry.BLOCKS.register(modEventBus);
+        ItemRegistry.ITEMS.register(modEventBus);
+        CreativeTabRegistry.CREATIVE_MODE_TABS.register(modEventBus);
+        NeoForge.EVENT_BUS.register(this);
+    }
+
+    private void commonSetup(final FMLCommonSetupEvent event)
+    {
+
+    }
+
+    @SubscribeEvent
+    public void onServerStarting(ServerStartingEvent event)
+    {
+
+    }
+
+    @EventBusSubscriber(modid = MOD_ID, bus = EventBusSubscriber.Bus.MOD, value = Dist.CLIENT)
+    public static class ClientModEvents
+    {
+        @SubscribeEvent
+        public static void onClientSetup(FMLClientSetupEvent event)
+        {
+
+        }
+    }
+}