+++ changelog.md
@@ -1,17 +1,49 @@
-![Stacked Blocks](https://cdn.modrinth.com/data/UYrNpKJN/images/28fbdff482a08cd94d0873f5bc26922efe5e5e11.png)
+![Stacked Blocks](https://i.imgur.com/8hNdx2J.png)
 
 ***
 
 ### ✨ What's New:
 
-- Leads can now be crafted with leather or slime balls, making stacked blocks easier to craft.
-
-![Lead Crafting Recipe](https://i.imgur.com/oHYveVA.gif)
-
-### 🛠️ Other Changes:
-
-- Blocks can no longer be made with rope from Farmer's Delight.
+- This initial release includes the following blocks for Farmer's Delight, which can be crafted with rope:
+    - Stacked Rich Soil
+  - Stacked Stone Blocks
+  - Stacked Cobblestone Blocks
+  - Stacked Netherrack Blocks
+  - Stacked Bricks
+  - Stacked Melons
+  - Stacked Pumpkins
+  - Stacked Lapis Blocks
+  - Stacked Redstone Blocks
+  - Stacked Coal Blocks
+  - Stacked Iron Blocks
+  - Stacked Gold Blocks
+  - Stacked Emerald Blocks
+  - Stacked Diamond Blocks
+  - Stacked Netherite Blocks
+  - Stacked Quartz Blocks
+  - Stacked Oak Logs
+  - Stacked Stripped Oak Logs
+  - Stacked Spruce Logs
+  - Stacked Stripped Spruce Logs
+  - Stacked Birch Logs
+  - Stacked Stripped Birch Logs
+  - Stacked Jungle Logs
+  - Stacked Stripped Jungle Logs
+  - Stacked Acacia Logs
+  - Stacked Stripped Acacia Logs
+  - Stacked Dark Oak Logs
+  - Stacked Stripped Dark Oak Logs
+  - Stacked Mangrove Logs
+  - Stacked Stripped Mangrove Logs
+  - Stacked Cherry Logs
+  - Stacked Stripped Cherry Logs
+  - Stacked Bamboo Blocks
+  - Stacked Stripped Bamboo Blocks
+  - Stacked Crimson Stems
+  - Stacked Stripped Crimson Stems
+  - Stacked Warped Stems
+  - Stacked Stripped Warped Stems
 
 ***
 
-Send your suggestions or report any issues on [GitHub](https://github.com/axperty/stackedblocks/issues/new). Support Stacked Blocks by [creating your own server](https://url-shortener.curseforge.com/RVyce) or [donating on PayPal](https://paypal.me/kevgelhorn). 
+Send your suggestions or report any issues on [GitHub](https://github.com/axperty/stackedblocks-farmersdelight/issues/new). Support Stacked Blocks: Farmer's Delight by [creating your own server](https://url-shortener.curseforge.com/RVyce) or [donating on PayPal](https://paypal.me/kevgelhorn).